+++ opentelemetry-sdk/src/opentelemetry/sdk/_metrics/_internal/instrument.py
@@ -15,7 +15,15 @@
 # pylint: disable=too-many-ancestors
 
 from logging import getLogger
-from typing import TYPE_CHECKING, Dict, Generator, Iterable, Optional, Union
+from typing import (
+    TYPE_CHECKING,
+    Dict,
+    Generator,
+    Iterable,
+    List,
+    Optional,
+    Union,
+)
 
 from opentelemetry._metrics import CallbackT
 from opentelemetry._metrics import Counter as APICounter
@@ -26,6 +34,7 @@ from opentelemetry._metrics import (
     ObservableUpDownCounter as APIObservableUpDownCounter,
 )
 from opentelemetry._metrics import UpDownCounter as APIUpDownCounter
+from opentelemetry._metrics._internal.instrument import CallbackOptions
 from opentelemetry.sdk._metrics.measurement import Measurement
 from opentelemetry.sdk.util.instrumentation import InstrumentationScope
 
@@ -93,7 +102,7 @@ class _Asynchronous:
         self._measurement_consumer = measurement_consumer
         super().__init__(name, callbacks, unit=unit, description=description)
 
-        self._callbacks = []
+        self._callbacks: List[CallbackT] = []
 
         if callbacks is not None:
 
@@ -101,24 +110,33 @@ class _Asynchronous:
 
                 if isinstance(callback, Generator):
 
-                    def inner(callback=callback) -> Iterable[Measurement]:
-                        return next(callback)
+                    # advance generator to it's first yield
+                    next(callback)
+
+                    def inner(
+                        options: CallbackOptions,
+                        callback=callback,
+                    ) -> Iterable[Measurement]:
+                        try:
+                            return callback.send(options)
+                        except StopIteration:
+                            return []
 
                     self._callbacks.append(inner)
                 else:
                     self._callbacks.append(callback)
 
-    def callback(self) -> Iterable[Measurement]:
+    def callback(
+        self, callback_options: CallbackOptions
+    ) -> Iterable[Measurement]:
         for callback in self._callbacks:
             try:
-                for api_measurement in callback():
+                for api_measurement in callback(callback_options):
                     yield Measurement(
                         api_measurement.value,
                         instrument=self,
                         attributes=api_measurement.attributes,
                     )
-            except StopIteration:
-                pass
             except Exception:  # pylint: disable=broad-except
                 _logger.exception(
                     "Callback failed for instrument %s.", self.name